# api/openapi/mint-server.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mint_server/agent/agent.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/user/avatar": {
      "put": {
        "operationId": "UserService_UpdateAvatar",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userUpdateAvatarResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userUpdateAvatarRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/v1/user/me": {
      "get": {
        "operationId": "UserService_GetMe",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userGetMeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/v1/user/nickname": {
      "put": {
        "operationId": "UserService_UpdateNickname",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userUpdateNicknameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userUpdateNicknameRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/v1/user/password": {
      "put": {
        "operationId": "UserService_UpdatePassword",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userUpdatePasswordResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userUpdatePasswordRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "mint_serverUserUserInfo": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "title": "用户ID"
        },
        "displayName": {
          "type": "string",
          "title": "用户展示名称"
        },
        "avatarUrl": {
          "type": "string",
          "title": "用户头像地址"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "title": "用户状态，1 表示启用"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "userGetMeResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "title": "业务状态码，0 表示成功"
        },
        "message": {
          "type": "string",
          "title": "业务响应消息"
        },
        "user": {
          "$ref": "#/definitions/mint_serverUserUserInfo",
          "title": "当前登录用户信息"
        }
      }
    },
    "userUpdateAvatarRequest": {
      "type": "object",
      "properties": {
        "avatarUrl": {
          "type": "string",
          "title": "新头像地址"
        }
      }
    },
    "userUpdateAvatarResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "title": "业务状态码，0 表示成功"
        },
        "message": {
          "type": "string",
          "title": "业务响应消息"
        },
        "user": {
          "$ref": "#/definitions/mint_serverUserUserInfo",
          "title": "更新后的用户信息"
        }
      }
    },
    "userUpdateNicknameRequest": {
      "type": "object",
      "properties": {
        "nickname": {
          "type": "string",
          "title": "新展示名称"
        }
      }
    },
    "userUpdateNicknameResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "title": "业务状态码，0 表示成功"
        },
        "message": {
          "type": "string",
          "title": "业务响应消息"
        },
        "user": {
          "$ref": "#/definitions/mint_serverUserUserInfo",
          "title": "更新后的用户信息"
        }
      }
    },
    "userUpdatePasswordRequest": {
      "type": "object",
      "properties": {
        "oldPassword": {
          "type": "string",
          "title": "当前密码"
        },
        "newPassword": {
          "type": "string",
          "title": "新密码"
        }
      }
    },
    "userUpdatePasswordResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "title": "业务状态码，0 表示成功"
        },
        "message": {
          "type": "string",
          "title": "业务响应消息"
        }
      }
    }
  }
}